updated package name to "extensions"
