+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/loadbalance/extensions/models/Split.java
@@ -16,20 +16,18 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.broker.loadbalance.extensible.models;
+package org.apache.pulsar.broker.loadbalance.extensions.models;
 
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 
 /**
- * Defines the information required to unload or transfer a service unit(e.g. bundle).
+ * Defines the information required for a bundle split.
  */
-public record Unload(String sourceBroker, String serviceUnit, Optional<String> destBroker) {
-    public Unload {
-        Objects.requireNonNull(sourceBroker);
-        Objects.requireNonNull(serviceUnit);
-    }
-    public Unload(String sourceBroker, String serviceUnit) {
-        this(sourceBroker, serviceUnit, Optional.empty());
+public record Split(String bundle, String sourceBroker, Map<String, Optional<String>> splitBundleToDestBroker) {
+
+    public Split {
+        Objects.requireNonNull(bundle);
     }
 }
