+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/loadbalance/extensions/models/Unload.java
@@ -16,4 +16,20 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.broker.loadbalance.extensible.models;
+package org.apache.pulsar.broker.loadbalance.extensions.models;
+
+import java.util.Objects;
+import java.util.Optional;
+
+/**
+ * Defines the information required to unload or transfer a service unit(e.g. bundle).
+ */
+public record Unload(String sourceBroker, String serviceUnit, Optional<String> destBroker) {
+    public Unload {
+        Objects.requireNonNull(sourceBroker);
+        Objects.requireNonNull(serviceUnit);
+    }
+    public Unload(String sourceBroker, String serviceUnit) {
+        this(sourceBroker, serviceUnit, Optional.empty());
+    }
+}
