+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/loadbalance/extensions/channel/models/UnloadTest.java
@@ -16,32 +16,38 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.broker.loadbalance.extensible.channel.models;
+package org.apache.pulsar.broker.loadbalance.extensions.channel.models;
 
 import static org.testng.Assert.assertEquals;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Optional;
-import org.apache.pulsar.broker.loadbalance.extensible.models.Split;
+import org.apache.pulsar.broker.loadbalance.extensions.models.Unload;
 import org.testng.annotations.Test;
 
 @Test(groups = "broker")
-public class SplitTest {
+public class UnloadTest {
 
     @Test
-    public void testConstructor() {
-        Map<String, Optional<String>> map = new HashMap<>();
-        map.put("C", Optional.of("test"));
+    public void testConstructors() {
 
-        Split split = new Split("A", "B", map);
-        assertEquals(split.bundle(), "A");
-        assertEquals(split.sourceBroker(), "B");
-        assertEquals(split.splitBundleToDestBroker(), map);
+        Unload unload1 = new Unload("A", "B");
+        assertEquals(unload1.sourceBroker(), "A");
+        assertEquals(unload1.serviceUnit(), "B");
+        assertEquals(unload1.destBroker(), Optional.empty());
+
+        Unload unload2 = new Unload("A", "B", Optional.of("C"));
+        assertEquals(unload2.sourceBroker(), "A");
+        assertEquals(unload2.serviceUnit(), "B");
+        assertEquals(unload2.destBroker(), Optional.of("C"));
     }
 
     @Test(expectedExceptions = NullPointerException.class)
-    public void testNullBundle() {
-        new Split(null, "A", Map.of());
+    public void testNullSourceBroker() {
+        new Unload(null, "A");
+    }
+
+    @Test(expectedExceptions = NullPointerException.class)
+    public void testNullServiceUnit() {
+        new Unload("A", null);
     }
 
 }
